수정: 외부채널 서비스 목적 별 모듈 생성
1. 외부 채널 모듈을 만들었다.
시작은 validator였다. validator가 다루고자하는 데이터 만큼 늘어날 것이라고 판단하여 모듈을 분리했다. 그런데, 저장하는 로직도 각 데이터 마다 특징이 있을 것이라고 판단하여 모듈을 따로 빼서 정의했다.
2. validator를 분리함
외부 채널에 대한 validator가 다루려고 하는 데이터 만큼 늘어날 것이라고 판단하여 각자 알아서 validation을 할 수 있도록 기본형의 validation 클래스를 만들었다.
