+++ app/services/external_channel/product/external_channel_product_saver.rb
@@ -1,0 +1,7 @@
+module ExternalChannel
+  module Product
+    class ExternalProductSaver
+
+    end
+  end
+end
